Initial activity buttons
Added buttons to main activity screen, one for taking a photo from
camera and one for browsing for an image.  Added strings for the
button messages so that it's easy to translate.

Added permissions/device feature requests in manifest so that it should be
able to use the camera, storage and location (for later on).

Actually implement camera button action: Send intent request to camera
so that we should receive a file name of the photo back. There are some
problems with this code which is largely based on tutorial, and it doesn't
use the response for anything but a message, but it does make photos.