Make edge detection work for now
Create an own layout for edge detection so that things like progress bar
and image view update actually work, and added a button to main activity
so that we can test it.

Cleaned up the code somewhat, code style etc.

The edge detection seems to be very slow for even the example picture,
so tested it with the app icon... We should find a way to speed this up, and
start edge detection before even selecting a path. So when we
combine this with the selection activity, we need to start the thread early.