Tuned edge detection even more
- Only write destination pixels once, so avoid a separate function call
  for each pixel
- Avoid clamping function call overhead by inlining the clamping
  procedure

The example (oranges) image now renders much quicker than before, but
it's still too slow for actual usage. On top of that it is entirely
black when the edge detection is done. It might be due to the fact that
the example image is a JPG image and the app icon is a PNG image, not
sure.