Readd bezier smoothening to adjusted path
We use Path.quadTo to get a nice quadratic bezier curve of the path,
just like we did while drawing the path before adjustment. However,
we did not use the smoothening to get nicer curves again.

Checked that this does not make the adjusted path any worse: it still
gets adjusted to the edges pretty well.