Allow non-external storage for camera
The documentation of getExternalStorageDirectory says that it does not
always give an "external" directory, giving the internal storage if
there is no external SD card, but this does not seem to be the case on
all devices. Therefore, fall back to internal storage in such a case.
Not sure if this location can be easily found again later on, but at
least it should work.

I cannot test it though (even if I force it to use the internal one)
since my internal storage is too full and currently don't have an
emulator ready.