Snap bitmap to view size if it was still larger and center bitmap
The bitmap was sometimes still larger after the scaling was done. This
seems to be mostly the case for landscape photos in landscape rotation.
Fixed this by making the height or width (whichever is the case) exactly
the view height or width and scale the other dimension again, before
actually creating the scaled bitmap.

However, now it would be placed at the left upper side of the screen,
which is not very aesthetic. Center the bitmap on the canvas, and make
sure that the drawn points, the shader and the adjusted points are correctly
placed on the bitmap. Some of this code should also help if we ever add
actual zooming (since then we need to know the relative positions as well).