Improve speed of path adjustment with geometrical hashing
During edge detection, create a grid of the bitmap, where
each cell in the grid is a bucket that is hashed using
its top left coordinate (which can be easily implemented
using a 2-dimensional array and division operators).
We place each edge point into its corresponding bucket.

Then, during path adjustment, we look in the bucket of
the current point and only compare distance with those
edges. If the minimal distance that we find this way
is higher than the distance to a border of the bucket,
then we have to look in that neighbor bucket as well
since there might be closer points.

This does not degrade the quality of the adjustment and
makes it almost real-time for even large photos.