Native implementation of options menu
Simon's version used a custom action bar that overlays the existing one,
which is considered kind of hacky. Instead, we now use the already
available action bar and use the options menu hook to add a native
option menu (note that three dotted button which is default for these
kinds of menus).

Also make sure to use the string references everywhere.