Use saved bundle instead of configuration changes
The problem with onConfigurationChanged is that it only works correctly
if the activity would simply be restarted for an orientation change. If
the app is stopped using the home button or anything, then the view is
actually forgotten and handling the config change will simply show
a black, broken view.

The android documentation also says that this method is a last-case
resort and you have to handle resources yourself (which we didn't). When
we use onSaveInstanceState, we do need to track resources in a Bundle,
but then it's explicit and sensible what is happening.

This might be slightly slower now since we only store a few number of
items from the view, not the whole source bitmap, and other objects need
to be restarted as well (say activity, view). But at least they restart
predictably.