Make sure path is adjusted if it was rotated during draw
This seems an awful lot of code to clean up a small bug that causes the
path to not get adjusted and the fill shader to not get applied when you
rotate while drawing a path at the same time. This does make it more
clear in what state we're in after threads have finished.