More memory management improvements during/after saving
When the selection object is finished (or even if it is
paused), we no longer need to have the bitmap information
(since a restart would just recreate it anyway). Therefore,
release the bitmap decoding memory as well as let the
garbage collector know that the bitmap is no longer used.