Moved memory release to destroy of activity
The instance saving and lifecycle activities somewhat interfere with
each other: onSaveInstanceState is not always called, so only do it in
onDestroy. That's also the only time we really want to do it, the
other moments (onPause and onStop) are nice to have (removes resources
when other applications are put in view) but not necessary to fix memory
problems.

Also moved some code to their own functions so we can call them
elsewhere (which is eventually not necessary) and for readability.