Added invert selection action button
When clicked, this button keeps the path but changes whether the
inside of the path or the outside (until the canvas boundaries) has the
"original" bitmap and which one has the filter applied. This makes it
possible to select an object, change the filter to colorize, for example,
and then invert the selection to get a colorized object and normal
colored background. In other words, we can easily apply the filter
on the object which gives interesting results and is easy to demonstrate.

Verified that the normal selection still works and the inverted version does
too (the only thing is that the line gets slightly broader when inverting,
but that looks nice since it shifts focus to the internal component).
The code itself is a bit more complex than I hoped (but I hoped it was a
one-line code change).