Snap to boundary of bitmap and fix neighbor distance calculation again
We want the boundary of the bitmap to behave like an edge as well.
This should be done in two cases:
- If we have a point that is already outside the bitmap, then it should
  be placed inside the bitmap, and stay there. We did not yet do the
  latter part.
- If we check a neighbor that is outside the bitmap, then we want to
  let the bitmap boundary act like the closest edge (at least if there
  is no closer edge than the boundary).

This was still somewhat problematic since we distances need to be
fixed so that they no longer go outside the bitmap (and thus we have
to check the minimal distance afterward), and the distances were
previously incorrect (used the block hash numbers instead of the
pixel coordinates, and the +blocksize for right/bottom neighbors
was needed after all). But the distances were also always positive,
which is not what we need (the squares take care of it anyway)
and made the top left corner annoying to find.

It should now work smoothly with most pictures, where it will often
take the bitmap boundary if it's close enough (and not too many noisy
edges around it), and hopefully the normal adjustments are now again
more like how it was before the introduction of geometrical hashing
(with less wonky lines).